Add buildCompleteWebsite.sh script
This script will clone/pull the xsf-tools and xeps repos, and will fetch
and build TeXML. Afterwards it uses xsf-tools to build all XHTML/PDF
output for the current XEPs, build a HTML table of the current XEPs
run pelican to generate our non-XEP website content and finally patch
the pelican output to include the generated HTML table of the current
XEPs.

Added content/extensions to .gitignore as it should only be filled with
the output of xsf-tools prior to running pelican.