Fixing the XMPP logo
Where the two arms of the logo cross, everything from the right-hand arm crosses under the left arm, with the exception of the darkest orange bit. That bit passes over the left arm. As the blue colored bit on the right arm again passes under, I feel that this is a a bug (if the blue had _also_ gone over, there would be an argument that the left hand arm goes _through_ the right hand arm - but it doesn't now).

This commit modifies the logo in such a way that all of the right hand arm goes under the left hand arm.