Replace favicons
This commit replaces the favicons used on the website. Unlike the old favicons, the new ones are
centered, and have a transparent background.

This commit replaces, but also removes, a lot of the older favicons. The new package is much
lighter, yet keeping the same compatibility level as before.

The package was generated using the SVG logo, on https://realfavicongenerator.net