Make inherit_sidebar do what I expect it to do
inherit_sidebar does currently not inherit the sidebar menu from the parent, or it
does not do it in a way that I expect it to do. This commit might be a bug fix, or
might be a result of me not understanding how to use it properly (if so, it should
probably not be merged).

This commit changes the behavior, to either use the sidebar in the same page (when
inherit_sidebar is not used), or use the sidebar of the direct ancestor (when
inherit_sidebar is used).