Please add txxmpp to clients.json
Due to missing perl libraries in newer Debian versions and other problems with sendxmpp I wrote new new tool named txxmpp.c based on the widely available strophe library.  We use it to push commit announcements to the gnupg-devel MUC.  It is a plain and simple tool with all instructions in the help output.  I am not sure whether it fits into the client list but I have not seen a better place, yet.