Force a redirect to compliance suite XEP
The target resource (xep-0443.html) is not served by the same nginx instance that serves the rest of the website (/extensions is managed and deployed separately as part of the xsf/xeps repo). Apparently nginx shortcuts and returns a 404 instead of a redirect when it can't find the target file.

Using a full https:// URL or appending the 'redirect' flag is enough to fix this. I opted for the latter.

This change is already tested and deployed 😇