/software/: Use words to describe compliance
The existing approach used a custom font to render some icons as the
only way of describing compliance level. For clients that don't load or
render custom remote fonts, this information is invisible. For clients
that do render fonts, the meaning of these icons is a mystery.

Additionally, for users with screen readers or similar assistive
technologies, these icons are also meaningless (especially since they
have no "alt").

This patch adds the actual words to describe the compliance level, so
that the information can be easily consumed by humans.