Explicitly use lua5.4
The binary "lua" might point to a different version. Given that this
specific version is currently used for CI builds, make sure the matching
version is used and document the same one as a dependency.

In particular, this avoids issues on distributions which make 5.1 the
default (since some modules are missing for it). This also makes it
clear that `lua5.4-expat` is the correct dependency local builds.