Infrastructure team membership updates (September 2023)
Per a vote on the iteam mailing list, the following changes have been
approved:

1) Add new members: Zash, jonas

Both have been practically active in iteam for a while, so this is just an
update to reflect that reality.

2) Remove Tobias

Tobias has been inactive for a while, and okayed his removal from the team
member listing.

3) Add singpolyma, root

Two new members are joining the team.