Remove link to unofficial Kik messenger repo
It's been reported that at https://xmpp.org/uses/instant-messaging/ there is a link to a repository owned by a GitHub account that is owned by a cracker group that uses the API to build spambots on the XMPP network.

I'm unsure of the affiliation of the repository owner. The link is of limited added value to the XSF website. To err on the side of caution, this commit removes that link.

fixes #1372