Review: The XMPP Newsletter September 2025 (#1565)
* Create delete.me

* Uploaded Gajim, Kaida and Convo imgs

* Uploaded 2025-10-05-newsletter.md

* Delete static/images/newsletter/2025-10-05/delete.me

* Update joined Videos and Talks sections

* Moved XMPP Public channels

Moved XMPP Public channels to the end after the XEP section.

* Update brasilian -> brazilian

* Fix brasilian to Brazilian

* Removed link to [Z3R0 Maverick-HQ](gemini://z3r0.maverick-hq.org/)

* Added python-nbxmpp 6.3.1 entry

* Update introductory paragrah 2025-10-05-newsletter.md

From:

Like this newsletter, many projects and their efforts in the XMPP community are a result of people’s voluntary work. If you are happy with the services and software you may be using, please consider saying thanks or help these projects! Interested in supporting the Newsletter team? Read more [at the bottom](#help-us-to-build-the-newsletter).


To:

The [XMPP Newsletter](/categories/newsletter/) is brought to you by the [XSF Communication Team](/about/xsf/comm-team/).

Just like any other product or project by the XSF, the Newsletter is the result of the voluntary work of its members and contributors. If you are happy with the services and software you may be using, please consider saying thanks or help these projects!

Interested in contributing to the [XSF Communication Team](/about/xsf/comm-team/)? Read more [at the bottom](#help-us-to-build-the-newsletter).

* Updated the list of Contributors in the English (original) entry

- Changed:
  - "General contributors" to "This month's contributors"
  - Cleaned and updated the list of contributors names to reflect the actual list of names from the pad.

* Deleted a duplicate entry for xmpp-dns

* Update Newsletter Contributors & Translations section

---------

Co-authored-by: Echolon <Echolon@users.noreply.github.com>