POC: Include contact information for XSF Members on the website
This is an example of how contact information from `members.json` could be used to render contact links, where members are presented on the website.

I know just enough of Hugo, HTML and CSS to be dangerous, but the approach taken here can (and should) be improved by someone who knows what they're doing. I'm committing this as a Proof of Concept. I'm hoping that it is useful for someone to copy/paste relevant bits, but basically redo it, in a proper way.

fixes #1572