Update XSF Board Member bios after 2025/2026 elections
Removes bios for Dave and Edward. Added one for Guus.

This still needs a bio for Michael.

Depending on how the tie-breaking process pans out, Arne-Bruen's bio might need to be replaced with one for Adrien.

The composition of Council has not changed after the elections (same candidates were voted in again).